Lots more stuff
This unfortunately includes a hack to fix an issue where the parser for
some reason refuses to finish parsing if the file ends with an
expression. Need to figure that one out, but for now a dummy function at
the end of the file will do the trick I suppose.