Started introducing an IL structure
Since currently we're just interpreting the AST more or less directly,
this could seem like overkill, but I want to compile to an IL that we
can interpret instead because there's a lot of things in that structure
we can optimize. I could of course just mutate the AST, but that feels a
lot ickier than creating a second structure that's meant specifically
for interpretation.