Globals seem to be done
At least the global table, which is basically complete. Currently the
compiler skips init expr's and array dimensions for these, as I plan to
build a "main" function that I will dump all these init expr's into (as
assignment statements) as well as all the top-level statements in the
program. This should simplify a bunch of stuff.