Main function AST building finished
Also added value.rs after move (seems I forgot that)

Also added ast::Statement::End, which really should be there anyways, I
just made it a ast::Node in the first place because of some parser
ambiguities I'm still not sure how to fix. The main problem is when
parsing a function decl, for instance, the parser will mistake "End
Function" as "End" followed by nothing, and fail to parse the
function. Need to come up with a way to fix that as well.