AAVBobs runs again
Unfortunately the speedup gained from all of this was only about 3x for
this effect (which is the only one that runs currently and has an FPS
counter, so it's my benchmark). I'm sure I can start doing smaller
optimizations though; there's definitely more gains to be had either.

Either way, though, functionality has caught up to the compiler
pre-optimization, so this branch is ready to be merged.